Made the email input field transparent and updated the placeholder text to display white text for easier visibility for the end user.
Updated the CSS to make a custom class for purple coloured buttons. Applied this class to the login button up the top (to clean up the code) and also applied it to the Submit button of the subscribe component.

Also changed the alignment of the components of the subscribe form to make it look more professional and aesthetically appeasing.